update profile and authService
